+++ react/components/pages/Addresses.tsx
@@ -3,41 +3,31 @@ import { graphql } from 'react-apollo'
 import { compose, branch, renderComponent, withProps } from 'recompose'
 import { FormattedMessage } from 'react-intl'
 import { EmptyState, Button } from 'vtex.styleguide'
-import { ContentWrapper } from 'vtex.my-account-commons'
 import { withRouter, Link } from 'vtex.my-account-commons/Router'
 
+import { withContentWrapper } from '../shared/withContentWrapper'
 import AddressesLoading from '../loaders/AddressesLoading'
 import AddressBox from '../Addresses/AddressBox'
 import Toast from '../shared/Toast'
 
 import GET_ADRESSES from '../../graphql/getAddresses.gql'
 
-export function headerConfig() {
-  const headerContent = (
+export const headerConfig = {
+  namespace: 'vtex-account__address-list',
+  titleId: 'pages.addresses',
+  headerContent: (
     <Link to="/addresses/new">
       <Button variation="primary" block size="small">
         <FormattedMessage id="addresses.addAddress" />
       </Button>
     </Link>
-  )
-
-  return {
-    namespace: 'vtex-account__address-list',
-    titleId: 'pages.addresses',
-    headerContent,
-  }
-}
-
-function renderWrapper(children: any) {
-  return <ContentWrapper {...headerConfig()}>{() => children}</ContentWrapper>
+  ),
 }
 
 function EmptyAddresses() {
   const title = <FormattedMessage id="addresses.notFound" />
 
-  const content = <EmptyState title={title} />
-
-  return renderWrapper(content)
+  return <EmptyState title={title} />
 }
 
 class Addresses extends Component<Props> {
@@ -59,7 +49,7 @@ class Addresses extends Component<Props> {
   }
 
   public render() {
-    const content = (
+    return (
       <div className="flex-ns flex-wrap-ns items-start-ns relative tl">
         {this.props.addresses.map(address => (
           <AddressBox
@@ -74,8 +64,6 @@ class Addresses extends Component<Props> {
         )}
       </div>
     )
-
-    return renderWrapper(content)
   }
 }
 
@@ -98,6 +86,7 @@ const enhance = compose<Props, void>(
     ({ data }: { data: Data }) => data.loading,
     renderComponent(AddressesLoading)
   ),
+  withContentWrapper(headerConfig),
   branch(
     ({ data }: { data: { profile: { addresses: Address[] } } }) =>
       data.profile == null || data.profile.addresses.length === 0,