+++ react/typings/vtex.address-form/helpers.d.ts
@@ -1,4 +1,5 @@
 declare module 'vtex.address-form/helpers' {
   export const addValidation
+  export const removeValidation
   export const injectRules
 }