+++ CHANGELOG.md
@@ -7,6 +7,11 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 
+### Fixed
+
+- Address editing when using geolocation.
+- Country change not affecting the address form rules.
+
 ## [0.29.0] - 2019-09-12
 
 ### Added